Improve search for non-AFDO images for performace comparison.
To compare O2 vs AFDO, we try to use images from the chrome-pfq.
We now search the PFQ image created by the previous version of
chromeos since that is the one that validated the chrome for the
current version. If we dont find a previous, we try 2 versions
before.

BUG=None
TEST=White testing.ran simplified version of buildbot_test_toolchains.

Change-Id: I800255508117c7aa7f91468243e4713432e5561a
Reviewed-on: https://chrome-internal-review.googlesource.com/249245
Commit-Ready: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>