[crosperf] Add results_report tests, minor refactors.
This adds unit tests for results_report.JSONResultsReport and
results_report.ParseChromeosImage. The test for the former aren't as
extensive as they could be, but I hope to add more in the future.

This also simplifies ParseChromeosImage, and does a few refactors to
JSONResultsReport.

This also fixes all of `cros lint`'s complaints about results_report.

BUG=None
TEST=./run_tests.sh

Change-Id: Ice4267a5ff2e10df6690595c6b80db1ee1977d50
Reviewed-on: https://chrome-internal-review.googlesource.com/267665
Commit-Ready: George Burgess <gbiv@google.com>
Tested-by: George Burgess <gbiv@google.com>
Reviewed-by: Caroline Tice <cmtice@google.com>