[crosperf] Change flags to cbuildbot.
The command for cbuildbot it buildbot_utils had hardcoded the
'--notests' options; this was right for the normal nightly jobs, but
not for the rotating tests.  This CL makes that option a variable, and
also allows for users to be able to add other custom flags to their
cbuildbot invocations.

BUG=chromium:631149
TEST=none

Change-Id: Ifd2abd42808d99a28c3f7860f67a2eb2c9cd0118
Reviewed-on: https://chrome-internal-review.googlesource.com/271005
Commit-Ready: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>
Reviewed-by: Yunlian Jiang <yunlian@google.com>