crosperf: refactor results_organizer; NFC.
As part of playing with results_report (so I could make it less
ChromeOS-dependent), I started refactoring results_organizer, which is
only used by results_report.

This was the result. The goal was to make it easier to understand what,
exactly, results_organizer is trying to do.

BUG=chromium:641098
TEST=./test_all.sh passes.

Change-Id: I8e53175fbc1ded6f23631550223905b3321376ca
Reviewed-on: https://chrome-internal-review.googlesource.com/280815
Commit-Ready: George Burgess <gbiv@google.com>
Tested-by: George Burgess <gbiv@google.com>
Reviewed-by: Caroline Tice <cmtice@google.com>