Fix test failures in results_report refactors.
...I have absolutely no clue how I missed these failures (since I had to
run results_report_unittest to run the new perf parser test), but
apparently I did. The failure happened because we had a _chart_js arg
name, which didn't match up with chart_js (in the caller).

This also fixes a few minor linter complaints that I introduced in
patch-set 2 of said review. All of the other files I touched are clean
(except for a complaint about use of _logger in one of them, but that
was there before I made my changes).

BUG=None
TEST=./run_tests.sh *actually* passes this time. I promise. :)

Change-Id: I98e43b29ff9cef5fc9cf33143adc26b98a861f3e
Reviewed-on: https://chrome-internal-review.googlesource.com/285436
Commit-Ready: George Burgess <gbiv@google.com>
Tested-by: George Burgess <gbiv@google.com>
Reviewed-by: Caroline Tice <cmtice@google.com>