user_activity: Added implementation for the metrics.
For every function, we compute a distance based on the sum of
inclusive value fraction differences of pairs of common child functions
from a sample, and reference function sets.

For every function, we assign a score based on the distance, inclusive
value fraction from the sample, and reference function sets.

For every Chrome OS component, we compute a set of metrics consisting
in the number of functions, the average, and cumulative distance and
the total score of the functions matching the group.

BUG=None
TEST=benchmark_metrics_unittest passes

Change-Id: I7810ec86bc4fa488b49ee18de3b3bda2ef95ba09
Reviewed-on: https://chrome-internal-review.googlesource.com/286161
Reviewed-by: Luis Lozano <llozano@chromium.org>
Tested-by: Evelina Dumitrescu <evelinad@google.com>
Reviewed-on: https://chromium-review.googlesource.com/435910
Commit-Ready: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>