[bisection tool] Don't make failure to flash bad image fatal error.
Currently during initial setup, if the known 'bad' image fails to
flash onto the device, the user is given the option to 'skip'.
This causes the entire bisection to fail/stop.  This CL changes
the skip to allow bisection to continue (it is not unreasonable for
a bad image to fail to flash).

BUG=chromium:641539
TEST=None

Change-Id: Iddc5312ecf593192d563766d677e8d401824ae7d
Reviewed-on: https://chrome-internal-review.googlesource.com/290617
Commit-Ready: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>
Reviewed-by: Yunlian Jiang <yunlian@google.com>
Reviewed-by: Ting-Yuan Huang <laszio@google.com>