[toolchain-utils] Update rotating testers to launch asynchronously.
Currently the rotating testers launch one trybot job, then wait for
it to finish (to get the trybot image name) before launching the
second job.  Now that all the testing happens directly in the
buildbot HWTest stages, this waiting in not needed.  This CL removes
the synchronous waiting from the rotating testers.  (Allows the two
jobs to run in parallel).

BUG=None
TEST=None

Change-Id: I6c96c684dc69b205b847b584b4bbc9e10cf5313f
Reviewed-on: https://chrome-internal-review.googlesource.com/302137
Commit-Ready: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>
Reviewed-by: Yunlian Jiang <yunlian@google.com>