[binary search tool] Allow non-flashing images to be fails.
Currently if an image is so bad that it can't be flashed, it gets
marked as 'skipped', which is different from bad/failed.  This can
cause the binary search to never find the bad image.  We need an
option to treat a non-flashing image as a failure.  That's what
this CL does.

BUG=None
TEST=Tested by Manoj: without this change, binary search failed to
find any bad packages, but with this change it worked.

Change-Id: Ib223338a1820314e85352827857fb8f3ba1ed73a
Reviewed-on: https://chrome-internal-review.googlesource.com/307824
Commit-Ready: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>
Reviewed-by: Manoj Gupta <manojgupta@google.com>