[toolchain-utils] Update waterfall report (email, archive, pruning).
The CL udpates the waterfall summary report script to: 1). Send an
email containing the reports to whoever ran the script; 2). Copy the
generated reports to
/google/data/rw/users/mo/mobiletc-prebuild/waterfall-reports
so everyone on the team can see all the reports;  3). Prunes failures
that are older than 7 days from the failure dictionary.

BUG=None
TEST=Tested in my own directory.

Change-Id: I624f2f6f7575bd1095969282e8cdcec4558b9191
Reviewed-on: https://chrome-internal-review.googlesource.com/308300
Commit-Ready: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>
Reviewed-by: Yunlian Jiang <yunlian@google.com>