Add argument options to waterfall summary report.
Now there are options to generate just the main waterfall report, or
just the rotating builders report, with or without the failures
section; or just the failures report section.  There's also the option
to NOT update the data files (makes it easier to re-run reports).  Also,
replaces the llvm rotating builder with the llvm-next rotating builder.

BUG=chromium:673833
TEST=Tested in my directory.

Change-Id: If784758112c36c3d18c4b292f9229b53b047848c
Reviewed-on: https://chrome-internal-review.googlesource.com/311775
Commit-Ready: Caroline Tice <cmtice@google.com>
Tested-by: Caroline Tice <cmtice@google.com>
Reviewed-by: Yunlian Jiang <yunlian@google.com>