[toolchain-utils] update script to generate waterfall reports.
This CL updates the script that generates the toolchain team
daily waterfall reports:  It adds an --email option, to specify
an alternate person to receive the report; it removes the waterfall
builders we no longer care about; and it doesn't attempt to update
report data when the URL file is inaccessible.  This is part of the
work to make these run automatically every day.

BUG=chromium:749962
TEST=Ran script in my directory and it worked.

Change-Id: Iaf7c763ee8b7e3f6d32fd617cf511713dcd892ce
Reviewed-on: https://chromium-review.googlesource.com/599005
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>