[toolchain-utils] Fix security issue.
Remove the chromeos-toolchain-credentials.json file, which should
not be in a publicly visible repo.  Update the buildbot_test_*
scripts to take an argument for the directory containing the
credentials file. Update the buildbot_utils.py file to use the
credentials file argument, and to default to the location in our
role account.

BUG=782452
TEST=With some code commented out, did basic testing of flag/options
to make sure they work as expected.

Change-Id: I47a98a69b9ba8d2704d25abc37fe4447fa6ba343
Reviewed-on: https://chromium-review.googlesource.com/757901
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>