[toolchain-utils] Update whitelists for debug_info_test stage.
The debug_info_test stage verifies that everything that is supposed
to be built with LLVM is actually being built with LLVM.  We have
to whitelist a few files/packages that must continue to be built
with GCC.  This CL updates the whitelists to make this stage pass
on AMD64.

BUG=chromium:817648
TEST= debug_info_test.py passed on my local terra build, with these
changes. DebugInfoTest stage passed on cros tryjob fizz-llvm-toolchain-tryjob.

Change-Id: Ia9c79b5f13663d1f8a0e2b948f03a532b46d270b
Reviewed-on: https://chromium-review.googlesource.com/964590
Commit-Ready: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Ting-Yuan Huang <laszio@chromium.org>