Crosperf: Treat benchmarks with story- test_args as different benchmarks
The bug refers to https://crbug.com/902466.

This patch modified the behavior of crosperf on determining duplicate
benchmarks. When there are two blocks with same benchmark name in
experiment file, we will append them to benchmark list first. After
that, we will rename the benchmark if it has test_args: 'story-filter'
or 'story-tag-filter'. The new benchmark name will be passed to
experiment so that we will treat them as two different runs and get
separate tables on the report.

BUG=chromium:902466
TEST=Passed all up-to-date unittests, tested and verified with one
experiment file containing following benchmark settings:
    benchmark: webrtc {
        iterations: 3
        test_args: --story-filter=datachannel
        suite: telemetry_Crosperf
    }
    benchmark: webrtc {
        iterations: 3
        test_args: --story-tag-filter=smoothness
        suite: telemetry_Crosperf
    }

Change-Id: Id733273a5f9f43d149407055c9c0da3b761ddeef
Reviewed-on: https://chromium-review.googlesource.com/1321415
Commit-Ready: Zhizhou Yang <zhizhouy@google.com>
Tested-by: Zhizhou Yang <zhizhouy@google.com>
Reviewed-by: Caroline Tice <cmtice@chromium.org>