toolchain-utils: add a top-level `yapf` file
yapf is a Python autoformatter, much like how clang-tidy is an
autoformatter for c/c++. It's available by default in the chroot, and
appears to be in use by our autotest friends.

I ran `yapf -i *.py`, and it only suggested a handful of changes from
our current style here. The only one I really noticed was that it
preferred

'foo %s' % (really_long_variable_name_one,
	    really_long_variable_name_two)

over

'foo %s' %
	(really_long_variable_name_one, really_long_variable_name_two)

...Which I think we do in some files, anyway.

This file magically applies to all `yapf` invocations inside of
toolchain-utils, so it's not something that people need to worry about.
Simply run `yapf -i file.py`, and it'll be automatically formatted to
conform to Chromium's style guidelines, which we try to follow here.

BUG=chromium:918755
TEST=`yapf *.py` left files mostly untouched. `cros lint` is just as
     (un)happy after the format as it was before.

Change-Id: Ie09694d8466b398ca07f5c8544597c9a28c7abfe
Reviewed-on: https://chromium-review.googlesource.com/1392445
Commit-Ready: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>