toolchain-utils: add a `cros lint` git hook
This adds a git hook to run `cros lint` on all of the files we're
committing. It also adds a wrapper around both check-lint and
check-format that runs the two in parallel without making the output
interleave. Better names for check-style are appreciated.

BUG=chromium:918755
TEST=Ran ./check-style on a few Python files. Got the expected
     output/lint errors/etc. Works both outside and inside of the
     chroot.
CQ-DEPEND=CL:1394285

Change-Id: I30c74662759f27abdfd663083d04002a05260bf7
Reviewed-on: https://chromium-review.googlesource.com/1395807
Commit-Ready: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>