git-hooks: support linting outside of the source tree
We appear to have workflows and committers that prefer to `push` from
outside of the CrOS source tree.

It's relatively straightforward to have fallbacks to support that, so I
don't see why not.

If you want to interact with this repository from outside of a CrOS
source tree, you now have two options:
- If you have a source tree somewhere on your machine, you can set
  CHROMEOS_ROOT_DIRECTORY to that path. If `check-lint` detects a usable
  `cros` in there, it'll function as usual.
- If you don't have a source tree on your machine, `check-lint` will
  fall back to pylint-only linting (with a warning). This isn't ideal
  (we miss shell linting, as well as a few other nits in Python,
  somehow), but should get us most of the way to happiness.

This also includes a refactor of `test` to `[[`

BUG=chromium:918755
TEST=Ran check-lint on some python files in a checkout outside and
     inside of the source tree

Change-Id: I101c4c8b142f0c8f78819779adcfd0b3b0e1d614
Reviewed-on: https://chromium-review.googlesource.com/1412682
Commit-Ready: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>