crosperf: Feature to ignore min and max value in results
This patch provides an option for user to ignore min and max value in
the results when generating report. User can use this feature by
specifying `ignore_min_max: True` (which by default is False) in
experiment file.

When values count in a single test is smaller than 3, we automatically
ignore this option for it, and print out a warning.

BUG=chromium:938758
TEST=Tested with examples in general/cwp mode report generation. Passed
all unit tests.

Change-Id: I36a4c4d99836c201cdd2f2f9f2a4b1a4ffdaa47d
Reviewed-on: https://chromium-review.googlesource.com/1521054
Commit-Ready: Zhizhou Yang <zhizhouy@google.com>
Tested-by: Zhizhou Yang <zhizhouy@google.com>
Reviewed-by: Zhizhou Yang <zhizhouy@google.com>