crosperf: Use --fast to prevent copying sysinfo for telemetry_Crosperf
We don't need sysinfo collected by perf from DUT. --fast will stop
copying sysinfo but was orignially disabled for profiling runs because
of a conflict in autotest profiler. Now that we don't use autotest
profiler any more, this patch will turn on --fast for telemetry_Crosperf
no matter profiler_args are present or not.

BUG=chromium:956174
TEST=Passed unit tests; Tested both with profiler_args and without,
checked the sysinfo directory and it was empty.

Change-Id: Id99db48d6ae9043309c5113d0a6d9afd6e3c095b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1583018
Reviewed-by: Luis Lozano <llozano@chromium.org>
Tested-by: Zhizhou Yang <zhizhouy@google.com>