rotating builders: Replace asuka with sentry, remove daisy_skate.
As we are moving towards using pool quota for hardware tests for our
rotating builders, we need to replace asuka with sentry, as the former
is not available in the pool and two boards are similar. Also remove
daisy_skate because it will be deprecated soon and also not available
in the pool.

BUG=chromium:948268
TEST=None

Change-Id: Ie83dae3fb9942e3834fbaa92002ecde0d0ccbf11
Reviewed-on: https://chromium-review.googlesource.com/1592428
Commit-Ready: Tiancong Wang <tcwang@google.com>
Tested-by: Tiancong Wang <tcwang@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>