clang_tidy: move scripts from sys-devel/llvm
This copies the scripts from sys-devel/llvm into toolchain-utils. For
reasons detailed on the bug, we believe that toolchain-utils is a better
home for these scripts than sys-devel/llvm.

Modulo formatting on files outside of clang-tidy-warn.py (please see the
patch after this in the relation chain for formats/updates/etc of that
:) ), this is a bit-for-bit copy of each of these files.

Removal of the copies of these in sys-devel/llvm (and updating ebuilds,
etc.) will be done once everything's in place in its new home, and once
we have the clang-tidy builder working with files provided by
toolchain-utils.

BUG=chromium:960495
TEST=None

Change-Id: I09a2fdb9db25059cf4fbf7e9516f7d4cacfd7bfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1600575
Reviewed-by: Caroline Tice <cmtice@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>