pgo_tools: make our profile uploader use hashes in file names
This CL makes us autodetect LLVM HEAD's hash from our PGO builders, and
use those in the names for our profiles.

Since it's cheap and otherwise a really subtle bug, this also adds a
safety check where we'll complain if not all LLVM HEADs from the user's
provided builds match. The way to override it is documented in
`die_with_head_complaint()` form.

BUG=None
TEST=Stubbed out _tar_and_upload_profdata and:
       - provided a bogus --llvm_hash
       - watched the script die when conflicting LLVM hashes were given
       - watched the script run under normal conditions

Change-Id: I26d17b6b698f14d40e7e11c55ef894afd91552f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1637657
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Zhizhou Yang <zhizhouy@google.com>