clang-tidy: Add warnings_pb2.py to clang-tidy work repo.
Ideally this file would be generated, via the compile_proto.sh.
However the Chrome OS builders do not have protc on them, and we
have no way of installing it on them, so we need to generate the
file ourselves and commit it (hence this CL).

BUG=chromium:973535
TEST=None

Change-Id: I7a738ed5268cf9e85c5c7953c3ffa890a7cfa740
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1656157
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Commit-Queue: Caroline Tice <cmtice@chromium.org>