Clang-Tidy: Add code search links to warnings HTML
This CL adds code search links to the HTML page generated by
clang_tidy_warn.py. When the code encounters a warning it does not know
how to build a link for, it will link to a query for the file in
chromium code search. This CL also adds the Chrome OS build version to the page

BUG=chromium:973560
TEST=All existing tests work

Change-Id: I9ecfa7e738b0c71958667a2898bac7f54427e490
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1659533
Reviewed-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Emma Vukelj <emmavukelj@google.com>