LLVM tools: patch management for LLVM and patch manager
The script 'llvm_patch_management.py' constructs the arguments for the
patch manager to handle patches of a package. The script 'patch_manager.py'
is the patch manager. The patch manager has different modes to handle
all applicable patches that fail to apply. The mode 'bisect_patches' is
a WIP.

BUG=None
TEST=Created a json file that has the patches of the LLVM ebuild and used
the default failure mode of the patch manager. All patches applied successfully
for r365631.

Change-Id: I42f339ac0bb7c35d70a96cc0c2daeb470336b13c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1708223
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Salud Lemus <saludlemus@google.com>