toolchain-utils: remove all xranges
This removes all mention of xrange from toolchain-utils (modulo ones
being changed in other CLs that are in flight). It's now an apparent
lint error to use xrange, and it hinders our move to python3.

As commented on If90d26664c70ccb73750f17573b89933fdb048f4, xrange ->
range in python2 is really only a space concern (or speed in
pathological cases), so migrations of this nature are generally super
straightforward. I glanced at each of these callsites, and none of them
appear to be pathological, so my hope is that this should all be Just
Fine :)

(Also fun to note that this includes a .diff file that has python code
embedded in it.)

BUG=None
TEST=Presubmit tests

Change-Id: Ic9f3ac3a5044d7a07da8a249bc505278d98203de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1717130
Commit-Queue: George Burgess <gbiv@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>