LLVM tools: Added support for bisection of patches
Changed 'get_google3_llvm_version.py' and 'get_llvm_hash.py' to use the
subprocess module instead of the command executer because the patch
manager script uses some methods for bisection (retrieving the good and
bad hash). They were also changed because the ebuilds will not have
access to the command executer when invoking the patch manager script.

BUG=None
TEST=Ran the script on SVN version of 367622 without continuing
bisection and successfully bisected the failed patch and terminated the
script. Then moved HEAD of source tree to the bisected SVN version and
applied the patch manually to test correctness and also moved HEAD to
the SVN version one above it and one below it and also manually applied
the patch.

Change-Id: I2367b76e9bb4d0c95d651ec7b3f66e582155d2c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1752526
Tested-by: Salud Lemus <saludlemus@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>