afdo: Add a metadata file for tracking latest AFDO artifacts.
As part of PUpr project, we want to move updates of AFDO profiles
out of PFQ. This is the first step to use a metadata file to track
the latest profiles for Chrome and kernel. An update to this file
will create a CL for updating Chrome ebuild or kernel ebuild.

BUG=chromium:990855
TEST=No test needed.

Change-Id: I3d1aad4e89cc6f09d6150bbd427ed07f42430597
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1752770
Reviewed-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Tiancong Wang <tcwang@google.com>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>