cros_utils: add an ExitStack utility
Python3 has a really handy tool called ExitStack. It's meant to make
context managers a bit more dynamic in what they can do. Sadly, Python2
doesn't have this.

This CL adds a simple implementation of ExitStack to our utilities. We
want this feature now, and the hope is that when we move to Python3,
we'll be able to just mass-replace contextlib3 with contextlib, and
everything will Just Work.

BUG=None
TEST=Unittest

Change-Id: I1e20271da598c86c00531821e1053bc33ecc2f63
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/toolchain-utils/+/1757218
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>